feat: extract common form wizard context to reuse in dupe form modal
