+++ frontend/src/features/workspace/components/CreateFormModal/CreateFormWizardContext.tsx
@@ -1,11 +1,9 @@
-import { createContext, useContext, useMemo, useState } from 'react'
-import { useForm, UseFormHandleSubmit, UseFormReturn } from 'react-hook-form'
+import { createContext, useContext } from 'react'
+import { UseFormHandleSubmit, UseFormReturn } from 'react-hook-form'
 
 import { FormResponseMode } from '~shared/types/form/form'
 
 import formsgSdk from '~utils/formSdk'
-
-import { useCreateFormMutations } from '~features/workspace/mutations'
 
 export enum CreateFormFlowStates {
   Landing = 'landing',
@@ -21,7 +19,7 @@ export type CreateFormWizardInputProps = {
   storageAck?: boolean
 }
 
-type CreateFormWizardContextReturn = {
+export type CreateFormWizardContextReturn = {
   currentStep: CreateFormFlowStates
   direction: number
   formMethods: UseFormReturn<CreateFormWizardInputProps>
@@ -34,91 +32,12 @@ type CreateFormWizardContextReturn = {
   handleBackToDetails: () => void
   keypair: ReturnType<typeof formsgSdk.crypto.generate>
   isLoading: boolean
+  modalHeader: string
 }
 
-const CreateFormWizardContext = createContext<
+export const CreateFormWizardContext = createContext<
   CreateFormWizardContextReturn | undefined
 >(undefined)
-
-const INITIAL_STEP_STATE: [CreateFormFlowStates, number] = [
-  CreateFormFlowStates.Details,
-  0 | 1 | -1,
-]
-
-const useCreateFormWizardContext = (): CreateFormWizardContextReturn => {
-  const [[currentStep, direction], setCurrentStep] =
-    useState(INITIAL_STEP_STATE)
-
-  /**
-   * Only used for storage mode forms, but generated first so that the key is
-   * immutable per open of the modal.
-   */
-  const keypair = useMemo(() => formsgSdk.crypto.generate(), [])
-
-  const formMethods = useForm<CreateFormWizardInputProps>({
-    defaultValues: {
-      responseMode: FormResponseMode.Encrypt,
-    },
-  })
-
-  const { handleSubmit } = formMethods
-
-  const { createEmailModeFormMutation, createStorageModeFormMutation } =
-    useCreateFormMutations()
-
-  const handleCreateStorageModeForm = handleSubmit(
-    ({ title, responseMode }) => {
-      if (responseMode !== FormResponseMode.Encrypt) return
-
-      return createStorageModeFormMutation.mutate({
-        title,
-        responseMode,
-        publicKey: keypair.publicKey,
-      })
-    },
-  )
-
-  const handleDetailsSubmit = handleSubmit((inputs) => {
-    if (inputs.responseMode === FormResponseMode.Email) {
-      return createEmailModeFormMutation.mutate({
-        emails: inputs.emails.filter(Boolean),
-        title: inputs.title,
-        responseMode: inputs.responseMode,
-      })
-    }
-    setCurrentStep([CreateFormFlowStates.Landing, 1])
-  })
-
-  const handleBackToDetails = () => {
-    setCurrentStep([CreateFormFlowStates.Details, -1])
-  }
-
-  return {
-    isLoading:
-      createEmailModeFormMutation.isLoading ||
-      createStorageModeFormMutation.isLoading,
-    keypair,
-    currentStep,
-    direction,
-    formMethods,
-    handleDetailsSubmit,
-    handleCreateStorageModeForm,
-    handleBackToDetails,
-  }
-}
-
-export const CreateFormWizardProvider = ({
-  children,
-}: {
-  children: React.ReactNode
-}): JSX.Element => {
-  const values = useCreateFormWizardContext()
-  return (
-    <CreateFormWizardContext.Provider value={values}>
-      {children}
-    </CreateFormWizardContext.Provider>
-  )
-}
 
 export const useCreateFormWizard = (): CreateFormWizardContextReturn => {
   const context = useContext(CreateFormWizardContext)
