+++ frontend/src/features/workspace/components/CreateFormModal/CreateFormWizardProvider.tsx
@@ -1,0 +1,117 @@
+import { useMemo, useState } from 'react'
+import { useForm } from 'react-hook-form'
+
+import { FormResponseMode } from '~shared/types'
+
+import formsgSdk from '~utils/formSdk'
+
+import { useCreateFormMutations } from '~features/workspace/mutations'
+
+import {
+  CreateFormFlowStates,
+  CreateFormWizardContext,
+  CreateFormWizardContextReturn,
+  CreateFormWizardInputProps,
+} from './CreateFormWizardContext'
+
+export const INITIAL_STEP_STATE: [CreateFormFlowStates, -1 | 1 | 0] = [
+  CreateFormFlowStates.Details,
+  -1,
+]
+
+interface UseCommonFormWizardProviderProps {
+  defaultValues?: Partial<CreateFormWizardInputProps>
+}
+
+export const useCommonFormWizardProvider = ({
+  defaultValues,
+}: UseCommonFormWizardProviderProps) => {
+  const [[currentStep, direction], setCurrentStep] =
+    useState(INITIAL_STEP_STATE)
+
+  /**
+   * Only used for storage mode forms, but generated first so that the key is
+   * immutable per open of the modal.
+   */
+  const keypair = useMemo(() => formsgSdk.crypto.generate(), [])
+
+  const formMethods = useForm<CreateFormWizardInputProps>({
+    defaultValues,
+  })
+
+  return {
+    formMethods,
+    keypair,
+    currentStep,
+    direction,
+    setCurrentStep,
+  }
+}
+
+const useCreateFormWizardContext = (): CreateFormWizardContextReturn => {
+  const { formMethods, currentStep, direction, keypair, setCurrentStep } =
+    useCommonFormWizardProvider({
+      defaultValues: {
+        responseMode: FormResponseMode.Encrypt,
+      },
+    })
+
+  const { handleSubmit } = formMethods
+
+  const { createEmailModeFormMutation, createStorageModeFormMutation } =
+    useCreateFormMutations()
+
+  const handleCreateStorageModeForm = handleSubmit(
+    ({ title, responseMode }) => {
+      if (responseMode !== FormResponseMode.Encrypt) return
+
+      return createStorageModeFormMutation.mutate({
+        title,
+        responseMode,
+        publicKey: keypair.publicKey,
+      })
+    },
+  )
+
+  const handleDetailsSubmit = handleSubmit((inputs) => {
+    if (inputs.responseMode === FormResponseMode.Email) {
+      return createEmailModeFormMutation.mutate({
+        emails: inputs.emails.filter(Boolean),
+        title: inputs.title,
+        responseMode: inputs.responseMode,
+      })
+    }
+    setCurrentStep([CreateFormFlowStates.Landing, 1])
+  })
+
+  const handleBackToDetails = () => {
+    setCurrentStep([CreateFormFlowStates.Details, -1])
+  }
+
+  return {
+    isLoading:
+      createEmailModeFormMutation.isLoading ||
+      createStorageModeFormMutation.isLoading,
+    keypair,
+    currentStep,
+    direction,
+    formMethods,
+    handleDetailsSubmit,
+    handleCreateStorageModeForm,
+    handleBackToDetails,
+    modalHeader: 'Set up your form',
+  }
+}
+
+export const CreateFormWizardProvider = ({
+  children,
+}: {
+  children: React.ReactNode
+}): JSX.Element => {
+  const values = useCreateFormWizardContext()
+  return (
+    <CreateFormWizardContext.Provider value={values}>
+      {children}
+    </CreateFormWizardContext.Provider>
+  )
+}
