+++ frontend/src/features/workspace/components/DuplicateFormModal/DupeFormWizardProvider.tsx
@@ -1,0 +1,78 @@
+import { FormResponseMode } from '~shared/types'
+
+import { useCreateFormMutations } from '~features/workspace/mutations'
+
+import {
+  CreateFormFlowStates,
+  CreateFormWizardContext,
+  CreateFormWizardContextReturn,
+} from '../CreateFormModal/CreateFormWizardContext'
+import { useCommonFormWizardProvider } from '../CreateFormModal/CreateFormWizardProvider'
+
+export const useDupeFormWizardContext = (): CreateFormWizardContextReturn => {
+  const { formMethods, currentStep, direction, keypair, setCurrentStep } =
+    useCommonFormWizardProvider({
+      defaultValues: {
+        responseMode: FormResponseMode.Encrypt,
+      },
+    })
+
+  const { handleSubmit } = formMethods
+
+  const { createEmailModeFormMutation, createStorageModeFormMutation } =
+    useCreateFormMutations()
+
+  const handleCreateStorageModeForm = handleSubmit(
+    ({ title, responseMode }) => {
+      if (responseMode !== FormResponseMode.Encrypt) return
+
+      return createStorageModeFormMutation.mutate({
+        title,
+        responseMode,
+        publicKey: keypair.publicKey,
+      })
+    },
+  )
+
+  const handleDetailsSubmit = handleSubmit((inputs) => {
+    if (inputs.responseMode === FormResponseMode.Email) {
+      return createEmailModeFormMutation.mutate({
+        emails: inputs.emails.filter(Boolean),
+        title: inputs.title,
+        responseMode: inputs.responseMode,
+      })
+    }
+    setCurrentStep([CreateFormFlowStates.Landing, 1])
+  })
+
+  const handleBackToDetails = () => {
+    setCurrentStep([CreateFormFlowStates.Details, -1])
+  }
+
+  return {
+    isLoading:
+      createEmailModeFormMutation.isLoading ||
+      createStorageModeFormMutation.isLoading,
+    keypair,
+    currentStep,
+    direction,
+    formMethods,
+    handleDetailsSubmit,
+    handleCreateStorageModeForm,
+    handleBackToDetails,
+    modalHeader: 'Duplicate form',
+  }
+}
+
+export const DupeFormWizardProvider = ({
+  children,
+}: {
+  children: React.ReactNode
+}): JSX.Element => {
+  const values = useDupeFormWizardContext()
+  return (
+    <CreateFormWizardContext.Provider value={values}>
+      {children}
+    </CreateFormWizardContext.Provider>
+  )
+}
