+++ frontend/src/features/workspace/components/DuplicateFormModal/index.ts
@@ -1,0 +1,1 @@
+export { DuplicateFormModal } from './DuplicateFormModal'
